--- faceADE-main-7925/faceADE.xcodeproj/xcuserdata/anthonynewman.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "FEEC16C8-849E-4517-BD5A-CD9A5D48DDFD"
-   type = "1"
-   version = "2.0">
-</Bucket>
