--- faceADE-main-7925/faceADE/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,14 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "faceADElogo3.png",
-      "idiom" : "universal",
-      "platform" : "ios",
-      "size" : "1024x1024"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
