--- faceADE-main-7925/faceADEUITests/faceADEUITestsLaunchTests.swift
@@ -1,32 +1,0 @@
-//
-//  faceADEUITestsLaunchTests.swift
-//  faceADEUITests
-//
-//  Created by Katherine Guo on 4/23/24.
-//
-
-import XCTest
-
-final class faceADEUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
